EvoNHI_MVP/pyproject.toml
[project]
name = "evonhi-saas"
version = "0.1.0"
description = "SaaS-first MVP for EvoNHI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.116.1",
    "uvicorn==0.35.0",
    "sqlalchemy==2.0.43",
    "psycopg[binary]==3.2.9",
    "pydantic==2.11.7",
    "networkx==3.5",
    "PyYAML==6.0.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
